+++ src/searchmob_desktop/engines/__init__.py
@@ -16,7 +16,7 @@ from searchmob_desktop.engines.marginalia import fetch_marginalia
 from searchmob_desktop.engines.mojeek import fetch_mojeek
 from searchmob_desktop.engines.mojeek_api import fetch_mojeek_api
 from searchmob_desktop.engines.mwmbl import fetch_mwmbl
-from searchmob_desktop.engines.normalize import normalize_url
+from searchmob_desktop.engines.normalize import normalize_url, strip_tracking_params
 from searchmob_desktop.engines.proxy import USER_AGENTS, make_privacy_client
 from searchmob_desktop.engines.types import EngineContext, SearchResult
 from searchmob_desktop.engines.wikipedia import fetch_wikipedia
@@ -38,4 +38,5 @@ __all__ = [
     "fetch_wikipedia",
     "make_privacy_client",
     "normalize_url",
+    "strip_tracking_params",
 ]